// types.ts
export type NotificationType = 'performance' | 'latency' | 'storage' | 'query' | 'load' | 'TABLE_SCAN' | 'JOIN_INEFFICIENCY' | 'WAREHOUSE_IDLE' | 'COST_SPIKE' | 'QUERY_ASSIGNED' | 'ASSIGNMENT_UPDATED';
export type NotificationSeverity = 'Info' | 'Warning' | 'Critical';

export interface Notification {
    id: string;
    insightTypeId: string;
    insightTopic: NotificationType;
    message: string;
    suggestions: string;
    timestamp: string; 
    warehouseName: string;
    queryId?: string;
    isRead: boolean;
    severity: NotificationSeverity;
}